+++ CHANGELOG.md
@@ -17,10 +17,12 @@ cases.
 
 ## [Unreleased]
 
+## [2.13.0] - 2024-08-30
+
 ### Added
 - Intel® VPL API 2.13 support, including new APIs for AV1 screen content tools,
-  encoded picture quality information, alpha channel encoding,
-  and Battlemage platform
+  encoded picture quality information, alpha channel encoding, and Battlemage
+  platform
 - hello-encode-jpeg example
 
 ## [2.12.0] - 2024-06-28
@@ -390,7 +392,8 @@ cases.
 - hello-transcode sample
 
 
-[Unreleased]: https://github.com/intel/libvpl/compare/v2.12.0...HEAD
+[Unreleased]: https://github.com/intel/libvpl/compare/v2.13.0...HEAD
+[2.13.0]: https://github.com/intel/libvpl/compare/v2.12.0...v2.13.0
 [2.12.0]: https://github.com/intel/libvpl/compare/v2.11.0...v2.12.0
 [2.11.0]: https://github.com/intel/libvpl/compare/v2.10.2...v2.11.0
 [2.10.2]: https://github.com/intel/libvpl/compare/v2.10.1...v2.10.2